+++ Domain-Module/Auth-Module/Auth-Application/src/main/java/com/pawith/authapplication/service/OAuthUseCase.java
@@ -5,6 +5,6 @@ import com.pawith.commonmodule.enums.Provider;
 
 public interface OAuthUseCase {
 
-    OAuthResponse oAuthLogin(Provider provider, String accessToken);
+    OAuthResponse oAuthLogin(Provider provider, String accessToken, String refreshToken);
 
 }
